refactor(tools): move README snippet lint
